core: Allow lists and maps as values in envvars
Using the environment variable methor for setting variable values, there
was previously no way to give a variable a value of a list or map. These
now support HCL for individual values - specifying:

    TF_VAR_test='["Hello", "World"]'

will set the variable `test` to a two-element list containing "Hello"
and "World". Note that care must be taken to not run into shell
expansion.